Design updates and a new event area on the website
The new event area is listing both new and past events. In the future we're going use these pages to promote events and let people apply for a ticket.

Only one button is now shown in the side bar. The other links is still there but with a new "text link" design.

The slideshow on the home page is paused for now.